fix(timestamp enlapsed): calculate enlapsed only from recv packet
Instead of recording myself the RTT, I am using the data received (ECHO_REPLY) by the packet
as the sent time.
The data received should be exactly the same as the one sent and I am sending the time
as the data.

It means that I don't have to keep track of the packet itself, I only print the difference
between the received time and the timestamp in the data field.