add(print ip/icmp packet): in case of error + verbose, print packets
Very long and complicated process to print packets.
Basically, the IP+ICMP are in the received packet data in case of an error,
because the sender wrapped the data (IP + ICMP sent) into a new IP packet.

Those header are long to print because of the format and the conversions as well.